Implement Match Timeline, Stats, and Players tabs; add match event components and bump minSdk to 26
Add a `HorizontalPager` to the Match screen to display a detailed Timeline, match Statistics, and Player information. Implement new UI components for match events, including points, game wins, and set wins, with platform-specific time formatting.

Key changes:
- Implement `TimelineList`, `Stats`, and `Players` components for match tracking
- Add `EventPoint`, `EventGameWin`, `EventSetWin`, and `EventDone` event visuals
- Bump `android-minSdk` to 26 to support modern `java.time` APIs
- Implement `ForcePortrait` orientation locking in the Navigator
- Refactor `coreModule` DI and update `RemoteMatch` data model nullability
- Enhance `CourtlyAppBar` and `SignOptionsScreen` with text overflow handling and UI refinements

+++ shared/src/commonMain/kotlin/com/mhd_07/courtly/feature_match_setup/domain/model/Setup.kt
@@ -15,7 +15,7 @@ data class Setup(
 
 //    val host : Player,
 
-    val type: MatchType = MatchType.Double,
+    val type: MatchType = MatchType.Single,
 
     val location: String = "",
